hybrid-array: rename `impls` module to `sizes`
It reflects what the module does: defines a macro for defining various
array sizes, then invokes it.
